SQL_Multi_agent_example_using_Autogen/requirements.txt
pyautogen
openai
sqlalchemy
pymysql
langchain
python-dotenv
pandas